moved managers
